pyproject.toml
[tool.poetry]
name = "pickpocket"
version = "0.1.0"
description = ""
authors = ["David Taylor <prooffreader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"
requests = "^2.22"
yagmail = "^0.11.224"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
